chore: split locator up into smaller files (#1238)
J=WAT-5650
TEST=manual

---------
