--- packages/visual-editor/src/components/Locator.tsx
@@ -1,2676 +1,0 @@
-import { FieldLabel, setDeep, WithPuckProps } from "@puckeditor/core";
-import {
-  FieldValueFilter,
-  FieldValueStaticFilter,
-  FilterSearchResponse,
-  Matcher,
-  NearFilterValue,
-  provideHeadless,
-  Result,
-  SearchHeadlessProvider,
-  SelectableStaticFilter,
-  useSearchActions,
-  useSearchState,
-} from "@yext/search-headless-react";
-import {
-  AnalyticsProvider,
-  AppliedFilters,
-  CardProps,
-  executeSearch,
-  Facets,
-  FilterSearch,
-  getUserLocation,
-  Coordinate,
-  MapboxMap,
-  MapMarkerOptions,
-  OnDragHandler,
-  OnSelectParams,
-  Pagination,
-  PinComponentProps,
-  SearchI18nextProvider,
-  useAnalytics as useSearchAnalytics,
-  VerticalResults,
-} from "@yext/search-ui-react";
-import React, { useEffect } from "react";
-import { useCollapse } from "react-collapsed";
-import { useTranslation } from "react-i18next";
-import {
-  FaChevronUp,
-  FaDotCircle,
-  FaRegCircle,
-  FaSlidersH,
-  FaTimes,
-} from "react-icons/fa";
-import {
-  type MultiSelectorOption,
-  type MultiSelectorValue,
-} from "../fields/MultiSelectorField.tsx";
-import { ImageField } from "../fields/ImageField.tsx";
-import { YextAutoField } from "../fields/YextAutoField.tsx";
-import { useDocument } from "../hooks/useDocument.tsx";
-import { usePreviewWindow } from "../hooks/usePreviewWindow.ts";
-import { getViewport, useWindowWidth } from "../hooks/useViewport.ts";
-import { Button } from "./atoms/button.tsx";
-import { ImageStylingFields } from "./contentBlocks/image/styling.ts";
-import { TranslatableString } from "../types/types.ts";
-import {
-  resolveLocalizedAssetImage,
-  TranslatableAssetImage,
-} from "../types/images.ts";
-import {
-  getPreferredDistanceUnit,
-  toMeters,
-  toMiles,
-} from "../utils/i18n/distance.ts";
-import { msg, pt } from "../utils/i18n/platform.ts";
-import { resolveComponentData } from "../utils/resolveComponentData.tsx";
-import {
-  createSearchAnalyticsConfig,
-  createSearchHeadlessConfig,
-} from "../utils/searchHeadlessConfig.ts";
-import { getThemeColorCssValue } from "../utils/colors.ts";
-import { ThemeColor, backgroundColors } from "../utils/themeConfigOptions.ts";
-import {
-  LocatorConfig,
-  StreamDocument,
-} from "../utils/types/StreamDocument.ts";
-import { getValueFromQueryString } from "../utils/urlQueryString.tsx";
-import { Body } from "./atoms/body.tsx";
-import { Heading } from "./atoms/heading.tsx";
-import {
-  DEFAULT_LOCATOR_RESULT_CARD_PROPS,
-  DistanceDisplayOption,
-  Location,
-  LocatorResultCard,
-  LocatorResultCardFields,
-  LocatorResultCardProps,
-} from "./LocatorResultCard.tsx";
-import { MapPinIcon } from "./MapPinIcon.js";
-import { useAnalytics } from "@yext/pages-components";
-import { useTemplateMetadata } from "../internal/hooks/useMessageReceivers.ts";
-import {
-  DEFAULT_ENTITY_TYPE,
-  LocatorEntityType,
-  isLocatorEntityType,
-  getLocatorEntityTypeSourceMap,
-  getEntityTypeLabel,
-} from "../utils/locatorEntityTypes.ts";
-import {
-  toPuckFields,
-  YextComponentConfig,
-  type YextCustomFieldRenderProps,
-  YextFields,
-} from "../fields/fields.ts";
-import { isVisualEditorTestEnv } from "./testing/utils.ts";
-
-const RESULTS_LIMIT = 20;
-const LOCATION_FIELD = "builtin.location";
-const COUNTRY_CODE_FIELD = "address.countryCode";
-const DEFAULT_MAP_CENTER: Coordinate = {
-  latitude: 40.741611,
-  longitude: -74.005371,
-}; // New York City
-const DEFAULT_RADIUS = 25;
-const HOURS_FIELD = "builtin.hours";
-const INITIAL_LOCATION_KEY = "initialLocation";
-const DEFAULT_TITLE = "Find a Location";
-const DEFAULT_DISTANCE_DISPLAY = "distanceFromUser";
-const DEFAULT_LOCATION_STYLE = {
-  pinIcon: { type: "none" },
-  pinColor: backgroundColors.background6.value,
-};
-const DEFAULT_PIN_ICON_WIDTH = 14;
-const MAX_PIN_ICON_WIDTH = 27;
-const PIN_ICON_MAX_FILE_SIZE_BYTES = 128 * 1024;
-const BOOLEAN_SUPPORTED_FIELDS = [
-  "primaryHeading",
-  "secondaryHeading",
-  "tertiaryHeading",
-] as const;
-
-type LocationStyleConfig = Record<
-  string,
-  {
-    color?: ThemeColor;
-    icon?: string;
-    customImage?: {
-      url: string;
-      width?: number;
-      aspectRatio?: number;
-    };
-  }
->;
-
-const LOCATOR_PIN_ICON_FIELD: ImageField = {
-  type: "image",
-  label: msg("fields.icon", "Icon"),
-  hideAltTextField: true,
-  maxFileSizeBytes: PIN_ICON_MAX_FILE_SIZE_BYTES,
-};
-
-const getConfiguredMapCenterOrDefault = (mapStartingLocation?: {
-  latitude: string;
-  longitude: string;
-}): Coordinate => {
-  if (mapStartingLocation?.latitude && mapStartingLocation.longitude) {
-    try {
-      return parseMapStartingLocation(mapStartingLocation);
-    } catch (e) {
-      console.error(e);
-    }
-  }
-
-  return DEFAULT_MAP_CENTER;
-};
-
-const getLocatorConfigFromPageSet = (pageSet?: string): LocatorConfig => {
-  if (!pageSet) {
-    return {};
-  }
-
-  try {
-    return JSON.parse(pageSet)?.typeConfig?.locatorConfig ?? {};
-  } catch {
-    console.error("Failed to parse locator config from page set");
-    return {};
-  }
-};
-
-const translateDistanceUnit = (
-  t: (key: string, options?: Record<string, unknown>) => string,
-  unit: "mile" | "kilometer",
-  count: number
-) => {
-  if (unit === "mile") {
-    return t("mile", { count, defaultValue: "mile" });
-  }
-
-  return t("kilometer", { count, defaultValue: "kilometer" });
-};
-
-const makiIconModules = import.meta.glob(
-  "../../node_modules/@mapbox/maki/icons/*.svg",
-  {
-    eager: true,
-    import: "default",
-  }
-) as Record<string, string>;
-
-const makiIconEntries = Object.entries(makiIconModules).map(([path, icon]) => {
-  const name = path.split("/").pop()?.replace(".svg", "") || path;
-  return [name, icon];
-});
-
-const makiIconMap: Record<string, string> = Object.fromEntries(makiIconEntries);
-
-const formatMakiIconLabel = (name: string) =>
-  name.replace(/[-_]/g, " ").replace(/\b\w/g, (char) => char.toUpperCase());
-
-const makiIconOptions = makiIconEntries.map(([name, icon]) => ({
-  label: formatMakiIconLabel(name),
-  value: name,
-  icon,
-}));
-
-const DEFAULT_MAKI_ICON_NAME = makiIconOptions[0]?.value;
-
-const ResultCardPropsField = ({
-  value,
-  onChange,
-}: {
-  value?: LocatorResultCardProps;
-  onChange: (value: LocatorResultCardProps) => void;
-}) => {
-  const streamDocument = useDocument();
-  const templateMetadata = useTemplateMetadata();
-  const entityTypeSourceMap = getLocatorEntityTypeSourceMap();
-  const entityTypeScopes = React.useMemo(() => {
-    const locatorConfig = getLocatorConfigFromPageSet(streamDocument?._pageset);
-    return locatorConfig.entityTypeScope ?? [];
-  }, [streamDocument]);
-
-  /**
-   * Builds the field schema for the result card editor, including:
-   * - Conditionally removing the primary CTA section when entity scope is not attached to a page set.
-   * - Toggling constant value vs. field selector visibility per section.
-   */
-  const resultCardFields = React.useMemo(() => {
-    if (!value?.entityType) {
-      return LocatorResultCardFields;
-    }
-    let fields = LocatorResultCardFields;
-    const entityTypeHasSourcePageSet = !!entityTypeSourceMap[value.entityType];
-    const scopeExistsForEntityType =
-      entityTypeScopes.find(
-        (scope) => scope.entityType === value.entityType
-      ) !== undefined;
-
-    fields = setDeep(
-      fields,
-      `objectFields.primaryCTA.objectFields.link.visible`,
-      !entityTypeHasSourcePageSet && scopeExistsForEntityType
-    );
-
-    // For each section, show either the field selector or the constant value editor.
-    BOOLEAN_SUPPORTED_FIELDS.forEach((key) => {
-      const headingConfig = value[key];
-      const constantValueEnabled = headingConfig?.constantValueEnabled ?? false;
-      const field = headingConfig?.field;
-      const fieldTypeId = field
-        ? templateMetadata?.locatorDisplayFields?.[field]?.field_type_id
-        : undefined;
-      const booleanFieldSelected =
-        !constantValueEnabled && fieldTypeId === "type.boolean";
-
-      fields = setDeep(
-        fields,
-        `objectFields.${key}.objectFields.field.visible`,
-        !constantValueEnabled
-      );
-      fields = setDeep(
-        fields,
-        `objectFields.${key}.objectFields.constantValue.visible`,
-        constantValueEnabled
-      );
-      fields = setDeep(
-        fields,
-        `objectFields.${key}.objectFields.trueDisplayText.visible`,
-        !constantValueEnabled && booleanFieldSelected
-      );
-      fields = setDeep(
-        fields,
-        `objectFields.${key}.objectFields.falseDisplayText.visible`,
-        booleanFieldSelected
-      );
-    });
-
-    const imageConstantValueEnabled =
-      value.image?.constantValueEnabled ?? false;
-    fields = setDeep(
-      fields,
-      "objectFields.image.objectFields.field.visible",
-      !imageConstantValueEnabled
-    );
-    fields = setDeep(
-      fields,
-      "objectFields.image.objectFields.constantValue.visible",
-      imageConstantValueEnabled
-    );
-
-    return fields;
-  }, [entityTypeSourceMap, entityTypeScopes, templateMetadata, value]);
-
-  return (
-    <YextAutoField
-      field={resultCardFields}
-      value={value ?? DEFAULT_LOCATOR_RESULT_CARD_PROPS}
-      onChange={onChange}
-    />
-  );
-};
-
-function getFacetFieldOptions(
-  entityTypes: LocatorEntityType[]
-): MultiSelectorOption<string>[] {
-  const facetFields: MultiSelectorOption<string>[] = [];
-  const addedValues: Set<string> = new Set<string>();
-  entityTypes.forEach((entityType) =>
-    getFacetFieldOptionsForEntityType(entityType).forEach((option) => {
-      if (option?.value && !addedValues.has(option.value)) {
-        facetFields.push(option);
-        addedValues.add(option.value);
-      }
-    })
-  );
-  return facetFields.sort((a, b) => a.label.localeCompare(b.label));
-}
-
-function getFacetFieldOptionsForEntityType(
-  entityType: LocatorEntityType
-): MultiSelectorOption<string>[] {
-  let filterOptions: MultiSelectorOption<string>[] = [
-    {
-      label: msg("fields.options.facets.city", "City"),
-      value: "address.city",
-    },
-    {
-      label: msg("fields.options.facets.postalCode", "Postal Code"),
-      value: "address.postalCode",
-    },
-    {
-      label: msg("fields.options.facets.region", "Region"),
-      value: "address.region",
-    },
-    {
-      label: msg("fields.options.facets.brandName", "Brand Name"),
-      value: "brandReference.name",
-    },
-  ];
-  switch (entityType) {
-    case "location":
-      filterOptions = filterOptions.concat(
-        {
-          label: msg("fields.options.facets.associations", "Associations"),
-          value: "associations",
-        },
-        {
-          label: msg("fields.options.facets.brands", "Brands"),
-          value: "brands",
-        },
-        {
-          label: msg("fields.options.facets.keywords", "Keywords"),
-          value: "keywords",
-        },
-        {
-          label: msg("fields.options.facets.languages", "Languages"),
-          value: "languages",
-        },
-        {
-          label: msg("fields.options.facets.paymentOptions", "Payment Options"),
-          value: "paymentOptions",
-        },
-        {
-          label: msg("fields.options.facets.products", "Products"),
-          value: "products",
-        },
-        {
-          label: msg("fields.options.facets.services", "Services"),
-          value: "services",
-        },
-        {
-          label: msg("fields.options.facets.specialties", "Specialties"),
-          value: "specialities",
-        }
-      );
-      break;
-    case "restaurant":
-      filterOptions = filterOptions.concat(
-        {
-          label: msg(
-            "fields.options.facets.acceptsReservations",
-            "Accepts Reservations"
-          ),
-          value: "acceptsReservations",
-        },
-        {
-          label: msg("fields.options.facets.associations", "Associations"),
-          value: "associations",
-        },
-        {
-          label: msg("fields.options.facets.brands", "Brands"),
-          value: "brands",
-        },
-        {
-          label: msg("fields.options.facets.keywords", "Keywords"),
-          value: "keywords",
-        },
-        {
-          label: msg("fields.options.facets.languages", "Languages"),
-          value: "languages",
-        },
-        {
-          label: msg("fields.options.facets.mealsServed", "Meals Served"),
-          value: "mealsServed",
-        },
-        {
-          label: msg("fields.options.facets.neighborhood", "Neighborhood"),
-          value: "neighborhood",
-        },
-        {
-          label: msg("fields.options.facets.paymentOptions", "Payment Options"),
-          value: "paymentOptions",
-        },
-        {
-          label: msg(
-            "fields.options.facets.pickupAndDeliveryServices",
-            "Pickup and Delivery Services"
-          ),
-          value: "pickupAndDeliveryServices",
-        },
-        {
-          label: msg("fields.options.facets.priceRange", "Price Range"),
-          value: "priceRange",
-        },
-        {
-          label: msg("fields.options.facets.services", "Services"),
-          value: "services",
-        },
-        {
-          label: msg("fields.options.facets.specialties", "Specialties"),
-          value: "specialities",
-        }
-      );
-      break;
-    case "healthcareFacility":
-      filterOptions = filterOptions.concat(
-        {
-          label: msg(
-            "fields.options.facets.acceptingNewPatients",
-            "Accepting New Patients"
-          ),
-          value: "acceptingNewPatients",
-        },
-        {
-          label: msg(
-            "fields.options.facets.conditionsTreated",
-            "Conditions Treated"
-          ),
-          value: "conditionsTreated",
-        },
-        {
-          label: msg(
-            "fields.options.facets.insuranceAccepted",
-            "Insurance Accepted"
-          ),
-          value: "insuranceAccepted",
-        },
-        {
-          label: msg("fields.options.facets.paymentOptions", "Payment Options"),
-          value: "paymentOptions",
-        },
-        {
-          label: msg("fields.options.facets.services", "Services"),
-          value: "services",
-        }
-      );
-      break;
-    case "healthcareProfessional":
-      filterOptions = filterOptions.concat(
-        {
-          label: msg(
-            "fields.options.facets.acceptingNewPatients",
-            "Accepting New Patients"
-          ),
-          value: "acceptingNewPatients",
-        },
-        {
-          label: msg(
-            "fields.options.facets.admittingHospitals",
-            "Admitting Hospitals"
-          ),
-          value: "admittingHospitals",
-        },
-        {
-          label: msg("fields.options.facets.brands", "Brands"),
-          value: "brands",
-        },
-        {
-          label: msg("fields.options.facets.certifications", "Certifications"),
-          value: "certifications",
-        },
-        {
-          label: msg(
-            "fields.options.facets.conditionsTreated",
-            "Conditions Treated"
-          ),
-          value: "conditionsTreated",
-        },
-        {
-          label: msg("fields.options.facets.degrees", "Degrees"),
-          value: "degrees",
-        },
-        {
-          label: msg("fields.options.facets.gender", "Gender"),
-          value: "gender",
-        },
-        {
-          label: msg(
-            "fields.options.facets.insuranceAccepted",
-            "Insurance Accepted"
-          ),
-          value: "insuranceAccepted",
-        },
-        {
-          label: msg("fields.options.facets.languages", "Languages"),
-          value: "languages",
-        },
-        {
-          label: msg("fields.options.facets.neighborhood", "Neighborhood"),
-          value: "neighborhood",
-        },
-        {
-          label: msg("fields.options.facets.officeName", "Office Name"),
-          value: "officeName",
-        },
-        {
-          label: msg("fields.options.facets.services", "Services"),
-          value: "services",
-        }
-      );
-      break;
-    case "hotel":
-      filterOptions = filterOptions.concat(
-        { label: msg("fields.options.facets.bar", "Bar"), value: "bar" },
-        {
-          label: msg("fields.options.facets.catsAllowed", "Cats Allowed"),
-          value: "catsAllowed",
-        },
-        {
-          label: msg("fields.options.facets.dogsAllowed", "Dogs Allowed"),
-          value: "dogsAllowed",
-        },
-        {
-          label: msg("fields.options.facets.parking", "Parking"),
-          value: "parking",
-        },
-        { label: msg("fields.options.facets.pools", "Pools"), value: "pools" }
-      );
-      break;
-    case "financialProfessional":
-      filterOptions = filterOptions.concat(
-        {
-          label: msg("fields.options.facets.certifications", "Certifications"),
-          value: "certifications",
-        },
-        {
-          label: msg("fields.options.facets.interests", "Interests"),
-          value: "interests",
-        },
-        {
-          label: msg("fields.options.facets.languages", "Languages"),
-          value: "languages",
-        },
-        {
-          label: msg("fields.options.facets.services", "Services"),
-          value: "services",
-        },
-        {
-          label: msg("fields.options.facets.specialties", "Specialties"),
-          value: "specialties",
-        },
-        {
-          label: msg(
-            "fields.options.facets.yearsOfExperience",
-            "Years of Experience"
-          ),
-          value: "yearsOfExperience",
-        }
-      );
-      break;
-    default:
-      break;
-  }
-  return filterOptions;
-}
-
-export interface LocatorProps {
-  /**
-   * The visual theme for the map tiles, chosen from a predefined list of Mapbox styles.
-   * @defaultValue 'mapbox://styles/mapbox/streets-v12'
-   */
-  mapStyle?: string;
-
-  /**
-   * Props to customize the locator map pin styles.
-   * Controls map pin appearance depending on the result's entity type.
-   * The number of entries is locked to the locator entity types for the page set.
-   */
-  locationStyles: Array<{
-    /** The entity type this style applies to. */
-    entityType: LocatorEntityType;
-    /** Whether to render an icon in the pin. */
-    pinIcon?: {
-      type: "none" | "icon" | "customImage";
-      /** Defaults to the first available Maki icon when type is 'icon'. */
-      iconName?: string;
-      /** Image rendered within the pin when type is 'customImage'. */
-      image?: TranslatableAssetImage;
-      /**
-       * Width of the custom image rendered within the pin.
-       * @defaultValue 14
-       * */
-      width?: number;
-      /** Aspect ratio of the custom image rendered within the pin. */
-      aspectRatio?: number;
-    };
-    /** The color applied to the pin. */
-    pinColor?: ThemeColor;
-  }>;
-
-  /**
-   * Configuration for the filters available in the locator search experience.
-   */
-  filters: {
-    /**
-     * If 'true', displays a button to filter for locations that are currently open.
-     * @defaultValue false
-     */
-    openNowButton: boolean;
-    /**
-     * If 'true', displays several distance options to filter searches to only locations within
-     * a certain radius.
-     * @defaultValue false
-     */
-    showDistanceOptions: boolean;
-    /** Accent color for filter button and icons. */
-    accentColor?: ThemeColor;
-    /** Which fields are facetable in the search experience */
-    facetFields?: MultiSelectorValue<string>;
-  };
-
-  /**
-   * The starting location for the map.
-   */
-  mapStartingLocation?: {
-    latitude: string;
-    longitude: string;
-  };
-  /**
-   * Configuration for the locator page heading.
-   * Allows customizing the title text and its color.
-   */
-  pageHeading?: {
-    /** The title displayed at the top of the locator page. */
-    title: TranslatableString;
-    /**
-     * The color applied to the locator page title.
-     * @defaultValue inherited from theme
-     */
-    color?: ThemeColor;
-  };
-  /**
-   * Props to customize the locator result card component.
-   * Controls which fields are displayed and their styling depending on the result's entity type.
-   * The number of entries is locked to the locator entity types for the page set.
-   */
-  resultCard: Array<{
-    /** Props to customize the locator result card component. */
-    props: LocatorResultCardProps;
-  }>;
-  /** Controls which distance value to display on each locator result card. */
-  distanceDisplay?: DistanceDisplayOption;
-}
-
-const locatorFields: YextFields<LocatorProps> = {
-  mapStyle: {
-    type: "basicSelector",
-    label: msg("fields.mapStyle", "Map Style"),
-    options: [
-      {
-        label: msg("fields.options.default", "Default"),
-        value: "mapbox://styles/mapbox/streets-v12",
-      },
-      {
-        label: msg("fields.options.satellite", "Satellite"),
-        value: "mapbox://styles/mapbox/satellite-streets-v12",
-      },
-      {
-        label: msg("fields.options.light", "Light"),
-        value: "mapbox://styles/mapbox/light-v11",
-      },
-      {
-        label: msg("fields.options.dark", "Dark"),
-        value: "mapbox://styles/mapbox/dark-v11",
-      },
-      {
-        label: msg("fields.options.navigationDay", "Navigation (Day)"),
-        value: "mapbox://styles/mapbox/navigation-day-v1",
-      },
-      {
-        label: msg("fields.options.navigationNight", "Navigation (Night)"),
-        value: "mapbox://styles/mapbox/navigation-night-v1",
-      },
-    ],
-  },
-  locationStyles: {
-    type: "array",
-    label: msg("fields.pinStyles", "Location styles"),
-    getItemSummary: (item: LocatorProps["locationStyles"][number]) =>
-      getEntityTypeLabel(item.entityType),
-    arrayFields: {
-      entityType: {
-        label: msg("fields.entityType", "Entity Type"),
-        type: "text",
-        visible: false,
-      },
-      pinIcon: {
-        type: "custom",
-        render: ({
-          value,
-          onChange,
-        }: YextCustomFieldRenderProps<
-          LocatorProps["locationStyles"][number]["pinIcon"]
-        >) => {
-          const selectedType = value?.type ?? "none";
-          return (
-            <div className="flex flex-col gap-3">
-              <YextAutoField
-                field={{
-                  type: "basicSelector",
-                  label: msg("fields.pinIcon", "Pin Icon"),
-                  options: [
-                    {
-                      label: msg("fields.options.none", "None"),
-                      value: "none",
-                    },
-                    {
-                      label: msg("fields.options.icon", "Icon"),
-                      value: "icon",
-                    },
-                    {
-                      label: msg("fields.options.customImage", "Custom image"),
-                      value: "customImage",
-                    },
-                  ],
-                }}
-                value={selectedType}
-                onChange={(type) =>
-                  onChange({
-                    ...value,
-                    type,
-                    iconName:
-                      type === "icon"
-                        ? (value?.iconName ?? DEFAULT_MAKI_ICON_NAME)
-                        : undefined,
-                  })
-                }
-              />
-              {selectedType === "icon" && (
-                <YextAutoField
-                  field={{
-                    type: "basicSelector",
-                    label: msg("fields.icon", "Icon"),
-                    options: makiIconOptions,
-                  }}
-                  value={value?.iconName}
-                  onChange={(iconName) =>
-                    onChange({ ...value, type: "icon", iconName })
-                  }
-                />
-              )}
-              {selectedType === "customImage" && (
-                <>
-                  <YextAutoField
-                    field={{
-                      ...LOCATOR_PIN_ICON_FIELD,
-                    }}
-                    value={value?.image}
-                    onChange={(image) =>
-                      onChange({ ...value, type: "customImage", image })
-                    }
-                  />
-                  <FieldLabel label={pt("fields.options.width", "Width")}>
-                    <YextAutoField
-                      field={{
-                        type: "number",
-                        min: 1,
-                        max: MAX_PIN_ICON_WIDTH,
-                      }}
-                      value={value?.width ?? DEFAULT_PIN_ICON_WIDTH}
-                      onChange={(width) =>
-                        onChange({
-                          ...value,
-                          type: "customImage",
-                          width,
-                        })
-                      }
-                    />
-                  </FieldLabel>
-                  <YextAutoField
-                    field={ImageStylingFields.aspectRatio}
-                    value={value?.aspectRatio}
-                    onChange={(aspectRatio) =>
-                      onChange({
-                        ...value,
-                        type: "customImage",
-                        aspectRatio,
-                      })
-                    }
-                  />
-                </>
-              )}
-            </div>
-          );
-        },
-      },
-      pinColor: {
-        type: "basicSelector",
-        label: msg("fields.pinColor", "Pin Color"),
-        options: "BACKGROUND_COLOR",
-      },
-    },
-    defaultItemProps: {
-      entityType: DEFAULT_ENTITY_TYPE,
-      pinIcon: { type: "none" },
-      pinColor: backgroundColors.background6.value,
-    },
-  },
-  filters: {
-    label: msg("fields.filters", "Filters"),
-    type: "object",
-    objectFields: {
-      openNowButton: {
-        label: msg("fields.options.includeOpenNow", "Include Open Now Button"),
-        type: "radio",
-        options: [
-          { label: msg("fields.options.yes", "Yes"), value: true },
-          { label: msg("fields.options.no", "No"), value: false },
-        ],
-      },
-      showDistanceOptions: {
-        label: msg(
-          "fields.options.showDistanceOptions",
-          "Include Distance Options"
-        ),
-        type: "radio",
-        options: [
-          { label: msg("fields.options.yes", "Yes"), value: true },
-          { label: msg("fields.options.no", "No"), value: false },
-        ],
-      },
-      accentColor: {
-        type: "basicSelector",
-        label: msg("fields.accentColor", "Accent Color"),
-        options: "SITE_COLOR",
-      },
-      facetFields: {
-        type: "multiSelector",
-        label: msg("fields.dynamicFilters", "Dynamic Filters"),
-        dropdownLabel: msg("fields.field", "Field"),
-        options: () => {
-          const entityTypeSourceMap = getLocatorEntityTypeSourceMap();
-          const entityTypes =
-            Object.keys(entityTypeSourceMap).filter(isLocatorEntityType);
-          return getFacetFieldOptions(entityTypes);
-        },
-        placeholderOptionLabel: msg(
-          "fields.options.selectAField",
-          "Select a field"
-        ),
-      } as any, // TODO(SUMO-8378): remove 'as any' when puck fixes objectFields typing
-    },
-  },
-  mapStartingLocation: {
-    type: "object",
-    label: msg("fields.options.mapStartingLocation", "Map Starting Location"),
-    objectFields: {
-      latitude: {
-        label: msg("fields.latitude", "Latitude"),
-        type: "text",
-      },
-      longitude: {
-        label: msg("fields.longitude", "Longitude"),
-        type: "text",
-      },
-    },
-  },
-  pageHeading: {
-    type: "object",
-    label: msg("fields.pageHeading", "Page Heading"),
-    objectFields: {
-      title: {
-        type: "translatableString",
-        label: msg("fields.title", "Title"),
-        filter: { types: ["type.string"] },
-      },
-      color: {
-        type: "basicSelector",
-        label: msg("fields.color", "Color"),
-        options: "SITE_COLOR",
-      },
-    },
-  },
-  resultCard: {
-    type: "array",
-    label: msg("fields.resultCard", "Result Card"),
-    getItemSummary: (item: LocatorProps["resultCard"][number]) =>
-      getEntityTypeLabel(item.props.entityType),
-    arrayFields: {
-      props: {
-        type: "custom",
-        render: ({
-          value,
-          onChange,
-        }: YextCustomFieldRenderProps<
-          LocatorProps["resultCard"][number]["props"]
-        >) => <ResultCardPropsField value={value} onChange={onChange} />,
-      },
-    },
-    defaultItemProps: {
-      props: DEFAULT_LOCATOR_RESULT_CARD_PROPS,
-    },
-  },
-  distanceDisplay: {
-    type: "basicSelector",
-    label: msg("fields.distanceDisplay", "Distance Display"),
-    options: [
-      {
-        label: msg("fields.options.distanceFromUser", "Distance from User"),
-        value: "distanceFromUser",
-      },
-      {
-        label: msg("fields.options.distanceFromSearch", "Distance from Search"),
-        value: "distanceFromSearch",
-      },
-      {
-        label: msg("fields.options.hidden", "Hidden"),
-        value: "hidden",
-      },
-    ],
-  },
-};
-
-/**
- * Available on Locator templates.
- */
-export const LocatorComponent: YextComponentConfig<LocatorProps> = {
-  fields: locatorFields,
-  /**
-   * Locks array lengths for `locationStyles` and `resultCard` to the current
-   * locator entity types so each entity type has exactly one entry.
-   */
-  resolveFields: (_data, params) => {
-    const entityDocument = params.metadata?.streamDocument;
-    const entityTypeSourceMap = entityDocument
-      ? getLocatorEntityTypeSourceMap(entityDocument)
-      : { [DEFAULT_ENTITY_TYPE]: undefined };
-    const entityTypes = Object.keys(
-      entityTypeSourceMap
-    ) as (keyof typeof entityTypeSourceMap)[];
-    const entityTypeCount = entityTypes.length;
-
-    let updatedFields: YextFields<LocatorProps> = { ...locatorFields };
-    updatedFields = setDeep(
-      updatedFields,
-      "locationStyles.min",
-      entityTypeCount
-    );
-    updatedFields = setDeep(
-      updatedFields,
-      "locationStyles.max",
-      entityTypeCount
-    );
-    updatedFields = setDeep(updatedFields, "resultCard.min", entityTypeCount);
-    updatedFields = setDeep(updatedFields, "resultCard.max", entityTypeCount);
-
-    return toPuckFields<LocatorProps>(updatedFields);
-  },
-  defaultProps: {
-    locationStyles: [],
-    resultCard: [],
-    filters: {
-      openNowButton: false,
-      showDistanceOptions: false,
-    },
-    pageHeading: {
-      title: { defaultValue: DEFAULT_TITLE },
-    },
-    distanceDisplay: DEFAULT_DISTANCE_DISPLAY,
-  },
-  label: msg("components.locator", "Locator"),
-  /**
-   * Reconciles `props.locationStyles` and `props.resultCard` so
-   * each list has exactly one entry per current locator entity type.
-   * Missing or mismatched entries are rebuilt from existing
-   * values and backfilled with defaults.
-   */
-  resolveData: (data, params) => {
-    const entityDocument = params.metadata?.streamDocument;
-    const entityTypeSourceMap = entityDocument
-      ? getLocatorEntityTypeSourceMap(entityDocument)
-      : { [DEFAULT_ENTITY_TYPE]: undefined };
-    const entityTypes = Object.keys(
-      entityTypeSourceMap
-    ) as (keyof typeof entityTypeSourceMap)[];
-
-    const previousLocationStyles = data.props.locationStyles ?? [];
-    const previousResultCard = data.props.resultCard ?? [];
-    const hasSameEntityTypes = (currentEntityTypes: string[]) =>
-      currentEntityTypes.length === entityTypes.length &&
-      entityTypes.every((entityType) =>
-        currentEntityTypes.includes(entityType)
-      );
-
-    const locationStylesByEntityType = new globalThis.Map(
-      previousLocationStyles
-        .filter((item) => !!item.entityType)
-        .map((item) => [item.entityType, item] as const)
-    );
-    const resultCardsByEntityType = new globalThis.Map(
-      previousResultCard
-        .filter((item) => !!item?.props?.entityType)
-        .map((item) => [item.props.entityType, item] as const)
-    );
-
-    const previouslocationStyleEntityTypes = previousLocationStyles.map(
-      (item) => item.entityType
-    );
-    const previousresultCardEntityTypes = previousResultCard.map(
-      (item) => item.props?.entityType
-    );
-
-    const shouldReconcileLocationStyles =
-      previousLocationStyles.length === 0 ||
-      !hasSameEntityTypes(previouslocationStyleEntityTypes);
-    const shouldReconcileResultCards =
-      previousResultCard.length === 0 ||
-      !hasSameEntityTypes(previousresultCardEntityTypes);
-
-    if (shouldReconcileLocationStyles) {
-      const newLocationStyles = entityTypes.map((entityType) => ({
-        ...DEFAULT_LOCATION_STYLE,
-        ...locationStylesByEntityType.get(entityType),
-        entityType,
-      }));
-      data = setDeep(data, "props.locationStyles", newLocationStyles);
-    }
-
-    if (shouldReconcileResultCards) {
-      const newResultCards = entityTypes.map((entityType) => {
-        const existing = resultCardsByEntityType.get(entityType);
-        return {
-          props: {
-            ...(existing?.props ?? DEFAULT_LOCATOR_RESULT_CARD_PROPS),
-            entityType,
-          },
-        };
-      });
-      data = setDeep(data, "props.resultCard", newResultCards);
-    }
-
-    return data;
-  },
-  render: (props) => <LocatorWrapper {...props} />,
-};
-
-const LocatorWrapper = (props: WithPuckProps<LocatorProps>) => {
-  const streamDocument = useDocument();
-  const { searchAnalyticsConfig, searcher } = React.useMemo(() => {
-    const searchHeadlessConfig = createSearchHeadlessConfig(
-      streamDocument,
-      props.puck.metadata?.experienceKeyEnvVar
-    );
-    if (searchHeadlessConfig === undefined) {
-      return { searchAnalyticsConfig: undefined, searcher: undefined };
-    }
-
-    const searchAnalyticsConfig = createSearchAnalyticsConfig(streamDocument);
-    return {
-      searchAnalyticsConfig,
-      searcher: provideHeadless(searchHeadlessConfig),
-    };
-  }, [streamDocument.id, streamDocument.locale]);
-
-  if (searcher === undefined || searchAnalyticsConfig === undefined) {
-    console.warn(
-      "Could not create Locator component because Search Headless or Search Analytics config is undefined. Please check your environment variables."
-    );
-    return <></>;
-  }
-  searcher.setSessionTrackingEnabled(true);
-  return (
-    <SearchHeadlessProvider searcher={searcher}>
-      <SearchI18nextProvider searcher={searcher}>
-        <AnalyticsProvider {...searchAnalyticsConfig}>
-          <LocatorInternal {...props} />
-        </AnalyticsProvider>
-      </SearchI18nextProvider>
-    </SearchHeadlessProvider>
-  );
-};
-
-type SearchState = "not started" | "loading" | "complete";
-
-const LoadingMapPlaceholder = () => {
-  const { t } = useTranslation();
-
-  return (
-    <div className="flex items-center justify-center w-full h-full">
-      <div className="border border-gray-300 rounded-lg p-6 bg-white shadow-md">
-        <Body className="text-gray-700" variant="lg">
-          {t("loadingMap", "Loading Map...")}
-        </Body>
-      </div>
-    </div>
-  );
-};
-
-const LocatorTestMap = () => {
-  return (
-    <div
-      className="relative h-full w-full overflow-hidden"
-      style={{
-        background:
-          "linear-gradient(135deg, #dcfce7 0%, #dbeafe 55%, #fde68a 100%)",
-      }}
-    >
-      <div
-        aria-hidden="true"
-        className="absolute inset-0 opacity-30"
-        style={{
-          backgroundImage:
-            "linear-gradient(rgba(255,255,255,0.2) 1px, transparent 1px), linear-gradient(90deg, rgba(255,255,255,0.2) 1px, transparent 1px)",
-          backgroundPosition: "0 0, 0 0",
-          backgroundSize: "48px 48px",
-        }}
-      />
-    </div>
-  );
-};
-
-const LocatorInternal = ({
-  mapStyle,
-  locationStyles,
-  filters: { openNowButton, showDistanceOptions, accentColor, facetFields },
-  mapStartingLocation,
-  resultCard: resultCardConfigs,
-  distanceDisplay,
-  pageHeading,
-}: LocatorProps) => {
-  // Adds unified [enable|disable]YextAnalytics to the window for both Pages and Search
-  // analytics. Typically used during consent banner implementation.
-  const searchAnalytics = useSearchAnalytics();
-  const pagesAnalytics = useAnalytics();
-  useEffect(() => {
-    (window as any).enableYextAnalytics = () => {
-      searchAnalytics?.optIn();
-      pagesAnalytics?.optIn();
-    };
-    (window as any).disableYextAnalytics = () => {
-      searchAnalytics?.optOut();
-      pagesAnalytics?.optOut();
-    };
-
-    return () => {
-      delete (window as any).enableYextAnalytics;
-      delete (window as any).disableYextAnalytics;
-    };
-  }, [searchAnalytics, pagesAnalytics]);
-
-  const { t, i18n } = useTranslation();
-  const previewWindow = usePreviewWindow();
-  const windowWidth = useWindowWidth(previewWindow);
-  const { isMobile } = getViewport(windowWidth);
-  const preferredUnit = getPreferredDistanceUnit(i18n.language);
-  const streamDocument = useDocument();
-  const entityTypeSourceMap = getLocatorEntityTypeSourceMap(streamDocument);
-  const entityTypes =
-    Object.keys(entityTypeSourceMap).filter(isLocatorEntityType);
-  const resultCount = useSearchState(
-    (state) => state.vertical.resultsCount || 0
-  );
-  const searchResults = useSearchState(
-    (state) => (state.vertical.results || []) as Result<Location>[]
-  );
-  const queryParamString =
-    typeof window === "undefined" ? "" : window.location.search;
-  const initialLocationParam = getValueFromQueryString(
-    INITIAL_LOCATION_KEY,
-    queryParamString
-  );
-
-  const iframe =
-    typeof document === "undefined"
-      ? undefined
-      : (document.getElementById("preview-frame") as HTMLIFrameElement);
-
-  let mapboxApiKey = streamDocument._env?.YEXT_MAPBOX_API_KEY;
-  if (
-    iframe?.contentDocument &&
-    streamDocument._env?.YEXT_EDIT_LAYOUT_MODE_MAPBOX_API_KEY
-  ) {
-    // If we are in the layout editor, use the non-URL-restricted Mapbox API key
-    mapboxApiKey = streamDocument._env.YEXT_EDIT_LAYOUT_MODE_MAPBOX_API_KEY;
-  }
-
-  const [showSearchAreaButton, setShowSearchAreaButton] = React.useState(false);
-  const [mapCenter, setMapCenter] = React.useState<Coordinate | undefined>();
-  const [mapRadius, setMapRadius] = React.useState<number | undefined>();
-  /** Explicit filter radius selected by the user, in meters */
-  const [selectedDistanceMeters, setSelectedDistanceMeters] = React.useState<
-    number | null
-  >(null);
-  const [selectedDistanceOption, setSelectedDistanceOption] = React.useState<
-    number | null
-  >(null);
-  /** Radius of last location near filter returned by the filter search API */
-  const apiFilterRadius = React.useRef<number | null>(null);
-
-  const handleDrag: OnDragHandler = (center, bounds) => {
-    setMapCenter({
-      latitude: center.latitude,
-      longitude: center.longitude,
-    });
-    setMapRadius(center.distanceTo(bounds.getNorthEast()));
-    setShowSearchAreaButton(true);
-  };
-
-  const [isOpenNowSelected, setIsOpenNowSelected] = React.useState(false);
-  const openNowFilter: SelectableStaticFilter = React.useMemo(
-    () => ({
-      filter: {
-        kind: "fieldValue",
-        fieldId: HOURS_FIELD,
-        matcher: Matcher.OpenAt,
-        value: "now",
-      },
-      selected: isOpenNowSelected,
-      displayName: t("openNow", "Open Now"),
-    }),
-    [isOpenNowSelected]
-  );
-
-  const handleSearchAreaClick = () => {
-    if (mapCenter && mapRadius) {
-      searchActions.setOffset(0);
-      const locationFilter: SelectableStaticFilter = {
-        selected: true,
-        displayName: "",
-        filter: {
-          kind: "fieldValue",
-          fieldId: "builtin.location",
-          value: {
-            lat: mapCenter.latitude,
-            lng: mapCenter.longitude,
-            radius: mapRadius,
-            name: t("customSearchArea", "Custom Search Area"),
-          },
-          matcher: Matcher.Near,
-        },
-      };
-      searchActions.setStaticFilters([locationFilter, openNowFilter]);
-      searchActions.executeVerticalQuery();
-      setSearchState("loading");
-      setShowSearchAreaButton(false);
-    }
-  };
-
-  const searchActions = useSearchActions();
-
-  const selectedFacets: string[] = React.useMemo(
-    () =>
-      facetFields?.selections
-        ?.filter((selection) => selection.value !== undefined)
-        ?.map((selection) => selection.value as string) ?? [],
-    [facetFields]
-  );
-  React.useEffect(() => {
-    searchActions.setFacetAllowList(selectedFacets);
-  }, [searchActions, selectedFacets]);
-
-  const filterDisplayName = useSearchState(
-    (state) =>
-      state.filters?.static?.find(
-        (filter) =>
-          filter.filter.kind === "fieldValue" &&
-          (filter.filter.fieldId === LOCATION_FIELD ||
-            filter.filter.fieldId === COUNTRY_CODE_FIELD)
-      )?.displayName
-  );
-  const handleFilterSelect = (params: OnSelectParams) => {
-    const newDisplayName = params.newDisplayName;
-    const filter = params.newFilter;
-
-    let locationFilter: SelectableStaticFilter;
-    let nearFilterValue: NearFilterValue | undefined;
-    switch (filter.matcher) {
-      case Matcher.Near: {
-        nearFilterValue = filter.value as NearFilterValue;
-        apiFilterRadius.current = nearFilterValue.radius;
-        // only overwrite radius from filter if display options are enabled
-        const radius =
-          showDistanceOptions && selectedDistanceMeters
-            ? selectedDistanceMeters
-            : nearFilterValue.radius;
-        locationFilter = buildNearLocationFilterFromPrevious(
-          nearFilterValue,
-          newDisplayName,
-          radius
-        );
-        break;
-      }
-      case Matcher.Equals: {
-        apiFilterRadius.current = null;
-        locationFilter = buildEqualsLocationFilter(filter, newDisplayName);
-        break;
-      }
-      default: {
-        throw new Error(`Unsupported matcher type: ${filter.matcher}`);
-      }
-    }
-
-    searchActions.setOffset(0);
-    searchActions.setStaticFilters([locationFilter, openNowFilter]);
-    searchActions.executeVerticalQuery();
-    setSearchState("loading");
-    if (
-      nearFilterValue?.lat &&
-      nearFilterValue?.lng &&
-      areValidCoordinates(nearFilterValue.lat, nearFilterValue.lng)
-    ) {
-      setMapCenter({
-        latitude: nearFilterValue.lat,
-        longitude: nearFilterValue.lng,
-      });
-      setMapRadius(nearFilterValue.radius);
-    }
-  };
-
-  const searchLoading = useSearchState((state) => state.searchStatus.isLoading);
-
-  const [searchState, setSearchState] =
-    React.useState<SearchState>("not started");
-
-  React.useEffect(() => {
-    if (!searchLoading && searchState === "loading") {
-      setSearchState("complete");
-    }
-  }, [searchLoading, searchState]);
-
-  React.useEffect(() => {
-    if (selectedDistanceOption === null) {
-      setSelectedDistanceMeters(null);
-      return;
-    }
-    setSelectedDistanceMeters(toMeters(selectedDistanceOption, preferredUnit));
-  }, [preferredUnit, selectedDistanceOption]);
-
-  const resultsRef = React.useRef<Array<HTMLDivElement | null>>([]);
-  const resultsContainer = React.useRef<HTMLDivElement>(null);
-  const [mobileResults, setMobileResults] = React.useState<Result<Location>[]>(
-    []
-  );
-  // Tracks the selected pin index to highlight the corresponding result card.
-  const [selectedResultIndex, setSelectedResultIndex] = React.useState<
-    number | null
-  >(null);
-
-  const setResultsRef = React.useCallback((index: number) => {
-    if (!resultsRef?.current) return null;
-    return (result: HTMLDivElement) => (resultsRef.current[index] = result);
-  }, []);
-
-  const scrollToResult = React.useCallback(
-    (result: Result<Location> | undefined) => {
-      if (result) {
-        if (typeof result.index === "number") {
-          setSelectedResultIndex(result.index);
-        }
-        let scrollPos = 0;
-        // the search results that are listed above this result
-        const previousResultsRef = resultsRef.current.filter(
-          (r, index) => r && result.index && index < result.index
-        );
-
-        // sum up the height of all search results that are listed above this result
-        if (previousResultsRef.length > 0) {
-          scrollPos = previousResultsRef
-            .map((elem) => elem?.scrollHeight ?? 0)
-            .reduce((total, height) => total + height);
-        }
-
-        resultsContainer.current?.scroll({
-          top: scrollPos,
-          behavior: "smooth",
-        });
-      } else {
-        setSelectedResultIndex(null);
-      }
-    },
-    [resultsContainer]
-  );
-
-  const markerOptionsOverride = React.useCallback(
-    (selected: boolean): MapMarkerOptions => {
-      return {
-        offset: (selected ? [0, -21] : [0, -14]) as [number, number],
-      };
-    },
-    []
-  );
-
-  const getResultCardProps = React.useCallback(
-    (entityType?: LocatorEntityType) => {
-      const existingConfig = (resultCardConfigs ?? []).find(
-        (item) => item.props.entityType === entityType
-      );
-      if (existingConfig) {
-        return existingConfig.props;
-      }
-      return DEFAULT_LOCATOR_RESULT_CARD_PROPS;
-    },
-    [resultCardConfigs]
-  );
-
-  const CardComponent = React.useCallback(
-    (result: CardProps<Location>) => {
-      let resultCardProps = DEFAULT_LOCATOR_RESULT_CARD_PROPS;
-      const resultEntityType = result.result.entityType;
-      if (resultEntityType && isLocatorEntityType(resultEntityType)) {
-        resultCardProps = getResultCardProps(resultEntityType);
-      } else {
-        console.warn(
-          "Unexpected entityType from search result: ",
-          resultEntityType
-        );
-      }
-      return (
-        <LocatorResultCard
-          {...result}
-          resultCardProps={resultCardProps}
-          isSelected={result.result.index === selectedResultIndex}
-          distanceDisplay={distanceDisplay}
-        />
-      );
-    },
-    [
-      distanceDisplay,
-      getResultCardProps,
-      entityTypeSourceMap,
-      selectedResultIndex,
-    ]
-  );
-
-  const [userLocationRetrieved, setUserLocationRetrieved] =
-    React.useState<boolean>(false);
-
-  const locationStylesConfig = React.useMemo(() => {
-    const config: LocationStyleConfig = {};
-    (locationStyles ?? []).forEach((locationStyle) => {
-      const entityType = locationStyle.entityType;
-      if (!entityType) return;
-      const iconValue =
-        locationStyle.pinIcon?.type === "icon"
-          ? locationStyle.pinIcon.iconName
-          : undefined;
-      const customImageValue =
-        locationStyle.pinIcon?.type === "customImage"
-          ? resolveLocalizedAssetImage(
-              locationStyle.pinIcon.image,
-              i18n.language
-            )
-          : undefined;
-      const customImageUrl = customImageValue?.url?.trim();
-      config[entityType] = {
-        color: locationStyle.pinColor,
-        icon:
-          typeof iconValue === "string" ? makiIconMap[iconValue] : undefined,
-        customImage: customImageUrl
-          ? {
-              url: customImageUrl,
-              width: locationStyle.pinIcon?.width,
-              aspectRatio: locationStyle.pinIcon?.aspectRatio,
-            }
-          : undefined,
-      };
-    });
-    return config;
-  }, [i18n.language, locationStyles]);
-
-  const initialMapCenter = React.useMemo(
-    () => getConfiguredMapCenterOrDefault(mapStartingLocation),
-    [mapStartingLocation]
-  );
-  const [centerCoords, setCenterCoords] =
-    React.useState<Coordinate>(initialMapCenter);
-  const [isInitialMapLocationResolved, setIsInitialMapLocationResolved] =
-    React.useState(false);
-  const canShowMoreMobileResults =
-    isMobile && mobileResults.length < resultCount;
-
-  const mapProps = React.useMemo(
-    () => ({
-      mapStyle,
-      centerCoords,
-      onDragHandler: handleDrag,
-      scrollToResult: scrollToResult,
-      markerOptionsOverride: markerOptionsOverride,
-      locationStyleConfig: locationStylesConfig,
-    }),
-    [
-      centerCoords,
-      handleDrag,
-      mapStyle,
-      markerOptionsOverride,
-      scrollToResult,
-      locationStylesConfig,
-    ]
-  );
-
-  React.useEffect(() => {
-    let isCancelled = false;
-
-    const resolveLocationAndSearch = async () => {
-      setIsInitialMapLocationResolved(false);
-      setCenterCoords(initialMapCenter);
-      setMapCenter(initialMapCenter);
-
-      const radius =
-        showDistanceOptions && selectedDistanceMeters
-          ? selectedDistanceMeters
-          : toMeters(DEFAULT_RADIUS, preferredUnit);
-      // default location filter to configured starting location or NYC
-      let initialLocationFilter = buildNearLocationFilterFromCoords(
-        initialMapCenter.latitude,
-        initialMapCenter.longitude,
-        radius
-      );
-      const doSearch = () => {
-        searchActions.setVerticalLimit(RESULTS_LIMIT);
-        searchActions.setOffset(0);
-        searchActions.setStaticFilters([initialLocationFilter]);
-        searchActions.executeVerticalQuery();
-        setSearchState("loading");
-        if (
-          initialLocationFilter.filter.kind === "fieldValue" &&
-          initialLocationFilter.filter.matcher === Matcher.Near
-        ) {
-          const filterValue = initialLocationFilter.filter
-            .value as NearFilterValue;
-          const centerCoords: Coordinate = {
-            longitude: filterValue.lng,
-            latitude: filterValue.lat,
-          };
-          if (
-            !isCancelled &&
-            areValidCoordinates(centerCoords.latitude, centerCoords.longitude)
-          ) {
-            setCenterCoords(centerCoords);
-            setMapCenter(centerCoords);
-            setMapRadius(filterValue.radius);
-          }
-        }
-        if (!isCancelled) {
-          setIsInitialMapLocationResolved(true);
-        }
-      };
-
-      const foundStartingLocationFromQueryParam = async (
-        queryParam: string
-      ): Promise<boolean> => {
-        return searchActions
-          .executeFilterSearch(queryParam, false, [
-            {
-              fieldApiName: LOCATION_FIELD,
-              entityType: entityTypes[0] ?? DEFAULT_ENTITY_TYPE,
-              fetchEntities: false,
-            },
-          ])
-          .then((response: FilterSearchResponse | undefined) => {
-            const firstResult = response?.sections[0]?.results[0];
-            const resultFilter = firstResult?.filter;
-            if (!firstResult || !resultFilter) {
-              return false;
-            }
-
-            switch (resultFilter.matcher) {
-              case Matcher.Near: {
-                const filterFromResult = resultFilter.value as NearFilterValue;
-                initialLocationFilter = buildNearLocationFilterFromPrevious(
-                  filterFromResult,
-                  firstResult.value
-                );
-                apiFilterRadius.current = filterFromResult.radius;
-                return true;
-              }
-              case Matcher.Equals: {
-                initialLocationFilter = buildEqualsLocationFilter(
-                  resultFilter,
-                  firstResult.value
-                );
-                apiFilterRadius.current = null;
-                return true;
-              }
-              default: {
-                return false;
-              }
-            }
-          })
-          .catch((e) => {
-            console.warn("Filter search for initial location failed:", e);
-            return false;
-          });
-      };
-
-      // 1. Check if a location could be determined from the initialLocation query parameter
-      if (
-        initialLocationParam &&
-        (await foundStartingLocationFromQueryParam(initialLocationParam))
-      ) {
-        doSearch();
-        return;
-      }
-
-      try {
-        // 2. Try to get user location via Geolocation API
-        const location = await getUserLocation();
-        const lat = location.coords.latitude;
-        const lng = location.coords.longitude;
-        setUserLocationRetrieved(true);
-
-        // Try to reverse-geocode the coordinates to a human-readable place name using Mapbox
-        let displayName: string | undefined;
-        try {
-          if (mapboxApiKey) {
-            const lang =
-              (streamDocument.locale as string) ||
-              (typeof navigator !== "undefined"
-                ? navigator.language
-                : undefined) ||
-              "en";
-            const url = `https://api.mapbox.com/geocoding/v5/mapbox.places/${lng},${lat}.json?access_token=${mapboxApiKey}&types=place,region,country&limit=1&language=${encodeURIComponent(
-              lang
-            )}`;
-
-            const res = await fetch(url);
-            if (res.ok) {
-              const data = await res.json();
-              const feature = data.features && data.features[0];
-              displayName = feature?.place_name || undefined;
-            }
-          }
-        } catch (e) {
-          console.warn("Reverse geocoding failed:", e);
-        } finally {
-          initialLocationFilter = buildNearLocationFilterFromCoords(
-            lat,
-            lng,
-            radius,
-            displayName
-          );
-        }
-      } catch {
-        // Fall back to the configured starting location or default center.
-      }
-
-      doSearch();
-    };
-
-    resolveLocationAndSearch().catch((e) => {
-      if (!isCancelled) {
-        setIsInitialMapLocationResolved(true);
-      }
-      console.error("Failed perform search:", e);
-    });
-    return () => {
-      isCancelled = true;
-    };
-  }, [initialLocationParam, initialMapCenter, searchActions]);
-
-  const handleOpenNowClick = (selected: boolean) => {
-    if (selected === isOpenNowSelected) {
-      // Prevents us from trying to set Open Now filter to false when it's not set
-      return;
-    }
-    searchActions.setFilterOption({
-      filter: {
-        kind: "fieldValue",
-        fieldId: HOURS_FIELD,
-        matcher: Matcher.OpenAt,
-        value: "now",
-      },
-      selected,
-      displayName: t("openNow", "Open Now"),
-    });
-    setIsOpenNowSelected(selected);
-    searchActions.setOffset(0);
-    executeSearch(searchActions);
-  };
-
-  const searchFilters = useSearchState((state) => state.filters);
-  const currentOffset = useSearchState((state) => state.vertical.offset);
-  const previousOffset = React.useRef<number | undefined>(undefined);
-  const prevIsMobile = React.useRef(isMobile);
-
-  // Scroll to top when pagination changes
-  React.useEffect(() => {
-    if (
-      !isMobile &&
-      currentOffset !== previousOffset.current &&
-      previousOffset.current !== undefined
-    ) {
-      resultsContainer.current?.scroll({
-        top: 0,
-        behavior: "smooth",
-      });
-    }
-    previousOffset.current = currentOffset;
-  }, [currentOffset, isMobile]);
-
-  React.useEffect(() => {
-    const switchedToMobile = isMobile && !prevIsMobile.current;
-
-    prevIsMobile.current = isMobile;
-
-    if (!switchedToMobile || searchLoading || (currentOffset ?? 0) === 0) {
-      return;
-    }
-
-    // Always reload from offset 0 if switching to mobile
-    setMobileResults([]);
-    searchActions.setOffset(0);
-    executeSearch(searchActions);
-    setSearchState("loading");
-  }, [currentOffset, isMobile, searchActions, searchLoading]);
-
-  React.useEffect(() => {
-    if (!isMobile || searchLoading) {
-      return;
-    }
-
-    setMobileResults((previousResults) => {
-      // Mobile keeps a single growing list: later offsets append, while
-      // offset 0 replaces the list after a fresh search.
-      return (currentOffset ?? 0) > 0 && searchResults.length > 0
-        ? [...previousResults, ...searchResults]
-        : searchResults;
-    });
-  }, [currentOffset, isMobile, searchLoading, searchResults]);
-
-  const handleDistanceClick = (
-    distance: number,
-    distanceUnit: "mile" | "kilometer"
-  ) => {
-    const existingFilters = searchFilters.static || [];
-    let updatedFilters: SelectableStaticFilter[];
-    const distanceInMeters = toMeters(distance, distanceUnit);
-    if (selectedDistanceOption === distance) {
-      setSelectedDistanceMeters(null);
-      setSelectedDistanceOption(null);
-      // revert to API radius (or default if none was found) if user clicks the same distance again
-      updatedFilters = updateRadiusInNearFiltersOnLocationField(
-        existingFilters,
-        apiFilterRadius.current ?? toMeters(DEFAULT_RADIUS, preferredUnit)
-      );
-    } else {
-      setSelectedDistanceOption(distance);
-      updatedFilters = updateRadiusInNearFiltersOnLocationField(
-        existingFilters,
-        distanceInMeters
-      );
-    }
-    searchActions.setStaticFilters(updatedFilters);
-    searchActions.setOffset(0);
-    executeSearch(searchActions);
-  };
-
-  const handleClearFiltersClick = () => {
-    const existingFilters = searchFilters.static || [];
-    // revert to API radius (or default if none was found)
-    const partiallyUpdatedFilters = updateRadiusInNearFiltersOnLocationField(
-      existingFilters,
-      apiFilterRadius.current ?? toMeters(DEFAULT_RADIUS, preferredUnit)
-    );
-    const updatedFilters = deselectOpenNowFilters(partiallyUpdatedFilters);
-
-    // Both open now and distance filters must be updated in the same setStaticFilters call to
-    // avoid problems due to the asynchronous nature of state updates.
-    searchActions.setStaticFilters(updatedFilters);
-    searchActions.resetFacets();
-    // Execute search to update AppliedFilters components
-    searchActions.setOffset(0);
-    executeSearch(searchActions);
-    setSelectedDistanceMeters(null);
-    setSelectedDistanceOption(null);
-  };
-
-  // If something else causes the filters to update, check if the hours filter is still present
-  // and toggle off the Open Now toggle if not.
-  React.useEffect(() => {
-    setIsOpenNowSelected(
-      searchFilters.static
-        ? !!searchFilters.static.find((staticFilter) => {
-            return (
-              staticFilter.filter.kind === "fieldValue" &&
-              staticFilter.filter.fieldId === HOURS_FIELD &&
-              staticFilter.selected === true
-            );
-          })
-        : false
-    );
-  }, [searchFilters]);
-
-  const hasFacetOptions =
-    (
-      useSearchState((state) =>
-        state.filters.facets?.filter((f) => f.options.length)
-      ) ?? []
-    ).length > 0;
-  const hasFilterModalToggle =
-    openNowButton || showDistanceOptions || hasFacetOptions;
-  const filterAccentColorCssVariable =
-    getThemeColorCssValue(accentColor?.selectedColor) ??
-    "var(--colors-palette-primary-dark)";
-  const [showFilterModal, setShowFilterModal] = React.useState(false);
-  const filterToggleButtonRef = React.useRef<HTMLButtonElement>(null);
-  const filterModalCloseButtonRef = React.useRef<HTMLButtonElement>(null);
-  const hasOpenedFilterModalRef = React.useRef(false);
-  const resolvedHeading =
-    (pageHeading?.title &&
-      resolveComponentData(pageHeading.title, i18n.language, streamDocument)) ||
-    t("findALocation", "Find a Location");
-
-  const requireMapOptIn: boolean = streamDocument.__?.visualEditorConfig
-    ? JSON.parse(streamDocument.__?.visualEditorConfig)?.requireMapOptIn
-    : false;
-  // If no opt-in is required, the map is already enabled.
-  const [mapEnabled, setMapEnabled] = React.useState(!requireMapOptIn);
-  // Adds unified [enable|disable]Map functions to the window.
-  useEffect(() => {
-    (window as any).enableMap = () => setMapEnabled(true);
-    (window as any).disableMap = () => setMapEnabled(false);
-
-    return () => {
-      delete (window as any).enableMap;
-      delete (window as any).disableMap;
-    };
-  }, []);
-
-  useEffect(() => {
-    if (showFilterModal) {
-      hasOpenedFilterModalRef.current = true;
-      filterModalCloseButtonRef.current?.focus();
-      return;
-    }
-
-    if (hasOpenedFilterModalRef.current) {
-      filterToggleButtonRef.current?.focus();
-    }
-  }, [showFilterModal]);
-
-  return (
-    <div className="components flex h-screen w-screen mx-auto">
-      {/* Left Section: FilterSearch + Results. Full width for small screens */}
-      <div
-        className="relative h-screen w-full md:w-2/5 lg:w-[40rem] flex flex-col md:min-w-[24rem]"
-        id="locatorLeftDiv"
-      >
-        <div className="px-8 py-6 gap-4 flex flex-col">
-          <Heading level={1} color={pageHeading?.color}>
-            {resolvedHeading}
-          </Heading>
-          <FilterSearch
-            searchFields={[
-              {
-                fieldApiName: LOCATION_FIELD,
-                entityType: entityTypes[0] ?? DEFAULT_ENTITY_TYPE, // only use a single entity type to avoid duplicate results
-              },
-            ]}
-            onSelect={(params) => handleFilterSelect(params)}
-            placeholder={t("searchHere", "Search here...")}
-            ariaLabel={t("findALocation", "Find a Location")}
-            customCssClasses={{
-              filterSearchContainer: "font-body-fontFamily",
-              focusedOption: "bg-gray-200 hover:bg-gray-200 block",
-              option: "hover:bg-gray-100 px-4 py-3",
-              inputElement:
-                "rounded-md p-4 h-11 font-body-fontFamily font-body-fontWeight text-body-fontSize placeholder:text-gray-700",
-              currentLocationButton:
-                "h-7 w-7 font-body-fontFamily font-body-fontWeight text-body-fontSize text-palette-primary-dark",
-              label:
-                "font-body-fontFamily font-body-fontWeight text-body-fontSize text-palette-primary-dark",
-            }}
-            showCurrentLocationButton={userLocationRetrieved}
-            geolocationProps={{
-              radius:
-                preferredUnit === "mile"
-                  ? DEFAULT_RADIUS
-                  : toMiles(DEFAULT_RADIUS), // this component uses miles, not meters
-            }}
-          />
-        </div>
-        <div className="relative flex-1 flex flex-col min-h-0">
-          <div className="px-8 py-4 text-body-fontSize border-y border-gray-300 inline-block">
-            <div className="flex flex-row justify-between" id="levelWithModal">
-              <ResultsCountSummary
-                searchState={searchState}
-                resultCount={resultCount}
-                selectedDistanceOption={selectedDistanceOption}
-                filterDisplayName={filterDisplayName}
-              />
-              {hasFilterModalToggle && (
-                <button
-                  ref={filterToggleButtonRef}
-                  className="inline-flex justify-between items-center gap-2 bg-white font-bold font-body-fontFamily text-body-sm-fontSize"
-                  style={{ color: filterAccentColorCssVariable }}
-                  onClick={() => setShowFilterModal((prev) => !prev)}
-                  aria-haspopup="dialog"
-                  aria-expanded={showFilterModal}
-                  aria-controls="locator-filter-modal"
-                >
-                  {t("filter", "Filter")}
-                  {<FaSlidersH />}
-                </button>
-              )}
-            </div>
-            <div className="flex flex-row justify-between">
-              <AppliedFilters
-                hiddenFields={[LOCATION_FIELD, COUNTRY_CODE_FIELD]}
-                customCssClasses={{
-                  removableFilter:
-                    "text-md font-normal mt-2 mb-0 font-body-fontFamily",
-                  clearAllButton: "hidden",
-                  appliedFiltersContainer: "mt-0 mb-0",
-                }}
-              />
-            </div>
-          </div>
-          {resultCount > 0 && (
-            <div
-              id="innerDiv"
-              className="md:flex-1 md:overflow-y-auto"
-              ref={resultsContainer}
-            >
-              {isMobile ? (
-                <MobileLocatorResultsSection
-                  CardComponent={CardComponent}
-                  results={mobileResults}
-                  hasMoreResults={canShowMoreMobileResults}
-                  handleShowMoreResults={() => {
-                    if (searchLoading || mobileResults.length >= resultCount) {
-                      return;
-                    }
-
-                    searchActions.setOffset(mobileResults.length);
-                    executeSearch(searchActions);
-                    setSearchState("loading");
-                  }}
-                />
-              ) : (
-                <VerticalResults
-                  CardComponent={CardComponent}
-                  setResultsRef={setResultsRef}
-                />
-              )}
-            </div>
-          )}
-          {!isMobile && resultCount > RESULTS_LIMIT && (
-            <div className="border-t border-gray-300 pt-4">
-              <Pagination
-                customCssClasses={{
-                  selectedLabel:
-                    "bg-palette-primary text-palette-primary-contrast border-palette-primary",
-                }}
-              />
-            </div>
-          )}
-          <FilterModal
-            showFilterModal={showFilterModal}
-            showOpenNowOption={openNowButton}
-            isOpenNowSelected={isOpenNowSelected}
-            handleOpenNowClick={handleOpenNowClick}
-            showDistanceOptions={showDistanceOptions}
-            selectedDistanceOption={selectedDistanceOption}
-            handleDistanceClick={handleDistanceClick}
-            handleCloseModalClick={() => setShowFilterModal(false)}
-            handleClearFiltersClick={handleClearFiltersClick}
-            accentColorCssValue={filterAccentColorCssVariable}
-            closeButtonRef={filterModalCloseButtonRef}
-          />
-        </div>
-      </div>
-
-      {/* Right Section: Map. Hidden for small screens */}
-      <div id="locatorMapDiv" className="md:flex-1 md:flex hidden relative">
-        {mapEnabled && isInitialMapLocationResolved && <Map {...mapProps} />}
-        {mapEnabled && !isInitialMapLocationResolved && (
-          <LoadingMapPlaceholder />
-        )}
-        {!mapEnabled && (
-          <div className="flex items-center justify-center w-full h-full bg-gray-100">
-            <div className="p-6">
-              <Body
-                className="text-gray-700 font-bold text-center"
-                variant="lg"
-              >
-                {t(
-                  "mapRequiresOptIn",
-                  "This map can only be displayed if cookies are enabled"
-                )}
-              </Body>
-              <div className="flex justify-center p-2">
-                <Button
-                  onClick={() => setMapEnabled(true)}
-                  className="py-2 px-4 basis-full sm:w-auto justify-center"
-                >
-                  {t("enableCookies", "Enable Cookies")}
-                </Button>
-              </div>
-            </div>
-          </div>
-        )}
-        {showSearchAreaButton && (
-          <div className="absolute top-10 left-0 right-0 flex justify-center">
-            <Button
-              onClick={handleSearchAreaClick}
-              className="py-2 px-4 shadow-xl"
-            >
-              <Body>{t("searchThisArea", "Search This Area")}</Body>
-            </Button>
-          </div>
-        )}
-      </div>
-    </div>
-  );
-};
-
-interface MobileLocatorResultsSectionProps {
-  CardComponent: React.ComponentType<CardProps<Location>>;
-  results: Result<Location>[];
-  hasMoreResults: boolean;
-  handleShowMoreResults: () => void;
-}
-
-const MobileLocatorResultsSection = ({
-  CardComponent,
-  results,
-  hasMoreResults,
-  handleShowMoreResults,
-}: MobileLocatorResultsSectionProps) => {
-  const { t } = useTranslation();
-
-  return (
-    <>
-      {results.length > 0 && (
-        <div>
-          {results.map((result, position) => (
-            <div
-              key={
-                result.rawData?.id ??
-                result.id ??
-                `${result.index ?? "result"}-${position}`
-              }
-            >
-              <CardComponent result={result} />
-            </div>
-          ))}
-        </div>
-      )}
-      {hasMoreResults && (
-        // Mobile replaces numbered pagination with incremental loading.
-        <div className="px-8 py-4">
-          <Button
-            className="w-full justify-center"
-            onClick={handleShowMoreResults}
-          >
-            {t("showMoreLocations", "Show more locations")}
-          </Button>
-        </div>
-      )}
-    </>
-  );
-};
-
-interface ResultsCountSummaryProps {
-  searchState: SearchState;
-  resultCount: number;
-  selectedDistanceOption: number | null;
-  filterDisplayName?: string;
-}
-
-const ResultsCountSummary = (props: ResultsCountSummaryProps) => {
-  const {
-    searchState,
-    resultCount,
-    selectedDistanceOption,
-    filterDisplayName,
-  } = props;
-  const { t, i18n } = useTranslation();
-
-  if (resultCount === 0) {
-    if (searchState === "not started") {
-      return (
-        <Body>
-          {t(
-            "useOurLocatorToFindALocationNearYou",
-            "Use our locator to find a location near you"
-          )}
-        </Body>
-      );
-    } else if (searchState === "complete") {
-      return (
-        <Body>
-          {t("noResultsFoundForThisArea", "No results found for this area")}
-        </Body>
-      );
-    } else {
-      return <div />;
-    }
-  } else {
-    if (filterDisplayName) {
-      if (selectedDistanceOption) {
-        const unit = getPreferredDistanceUnit(i18n.language);
-        return (
-          <Body>
-            {t("locationsWithinDistanceOf", {
-              count: resultCount,
-              distance: selectedDistanceOption,
-              unit: translateDistanceUnit(t, unit, selectedDistanceOption),
-              name: filterDisplayName,
-            })}
-          </Body>
-        );
-      } else {
-        return (
-          <Body>
-            {t("locationsNear", {
-              count: resultCount,
-              name: filterDisplayName,
-            })}
-          </Body>
-        );
-      }
-    } else {
-      return (
-        <Body>
-          {t("locationWithCount", {
-            count: resultCount,
-          })}
-        </Body>
-      );
-    }
-  }
-};
-
-interface MapProps {
-  mapStyle?: string;
-  centerCoords?: Coordinate;
-  onDragHandler?: OnDragHandler;
-  scrollToResult?: (result: Result<Location> | undefined) => void;
-  markerOptionsOverride?: (selected: boolean) => MapMarkerOptions;
-  locationStyleConfig?: LocationStyleConfig;
-}
-
-const Map: React.FC<MapProps> = ({
-  mapStyle,
-  centerCoords,
-  onDragHandler,
-  scrollToResult,
-  markerOptionsOverride,
-  locationStyleConfig,
-}) => {
-  const entityDocument: StreamDocument = useDocument();
-
-  const documentIsUndefined = typeof document === "undefined";
-  const iframe = documentIsUndefined
-    ? undefined
-    : (document.getElementById("preview-frame") as HTMLIFrameElement);
-
-  const locatorMapDiv = documentIsUndefined
-    ? null
-    : ((iframe?.contentDocument || document)?.getElementById(
-        "locatorMapDiv"
-      ) as HTMLDivElement | null);
-
-  const mapPadding = React.useMemo(
-    () => getMapboxMapPadding(locatorMapDiv),
-    [locatorMapDiv]
-  );
-  const mapboxOptions = React.useMemo(
-    () => ({
-      center: centerCoords,
-      fitBoundsOptions: { padding: mapPadding },
-      ...(mapStyle ? { style: mapStyle } : {}),
-    }),
-    [centerCoords, mapPadding, mapStyle]
-  );
-  const PinComponent = React.useMemo(
-    () =>
-      function PinComponent<T>(pinProps: PinComponentProps<T>) {
-        return (
-          <LocatorMapPin
-            {...pinProps}
-            locationStyleConfig={locationStyleConfig}
-          />
-        );
-      },
-    [locationStyleConfig]
-  );
-
-  if (isVisualEditorTestEnv()) {
-    return <LocatorTestMap />;
-  }
-
-  // During page generation we don't exist in a browser context
-  //@ts-expect-error MapboxGL is not loaded in the iframe content window
-  if (iframe?.contentDocument && !iframe.contentWindow?.mapboxgl) {
-    // We are in an iframe, and mapboxgl is not loaded in yet
-    return <LoadingMapPlaceholder />;
-  }
-
-  let mapboxApiKey = entityDocument._env?.YEXT_MAPBOX_API_KEY;
-  if (
-    iframe?.contentDocument &&
-    entityDocument._env?.YEXT_EDIT_LAYOUT_MODE_MAPBOX_API_KEY
-  ) {
-    // If we are in the layout editor, use the non-URL-restricted Mapbox API key
-    mapboxApiKey = entityDocument._env.YEXT_EDIT_LAYOUT_MODE_MAPBOX_API_KEY;
-  }
-
-  return (
-    <MapboxMap
-      mapboxAccessToken={mapboxApiKey || ""}
-      mapboxOptions={mapboxOptions}
-      onDrag={onDragHandler}
-      PinComponent={PinComponent}
-      iframeWindow={iframe?.contentWindow ?? undefined}
-      allowUpdates={!!iframe?.contentDocument}
-      onPinClick={scrollToResult}
-      markerOptionsOverride={markerOptionsOverride}
-    />
-  );
-};
-
-type LocatorMapPinProps<T> = PinComponentProps<T> & {
-  locationStyleConfig?: LocationStyleConfig;
-};
-
-const LocatorMapPin = <T,>(props: LocatorMapPinProps<T>) => {
-  const { result, selected, locationStyleConfig } = props;
-  const entityType = result.entityType;
-  const entityLocationStyle = entityType
-    ? locationStyleConfig?.[entityType]
-    : undefined;
-
-  return (
-    <MapPinIcon
-      color={entityLocationStyle?.color}
-      resultIndex={result.index}
-      icon={entityLocationStyle?.customImage?.url ?? entityLocationStyle?.icon}
-      iconWidth={
-        entityLocationStyle?.customImage?.width ?? DEFAULT_PIN_ICON_WIDTH
-      }
-      disableContrastIcon={!!entityLocationStyle?.customImage}
-      aspectRatio={entityLocationStyle?.customImage?.aspectRatio}
-      selected={selected}
-    />
-  );
-};
-
-interface FilterModalProps {
-  showFilterModal: boolean;
-  showOpenNowOption: boolean; // whether to show the Open Now filter option
-  isOpenNowSelected: boolean; // whether the Open Now filter is currently selected by the user
-  showDistanceOptions: boolean; // whether to show the Distance filter option
-  selectedDistanceOption: number | null;
-  handleCloseModalClick: () => void;
-  handleOpenNowClick: (selected: boolean) => void;
-  handleDistanceClick: (
-    distance: number,
-    distanceUnit: "mile" | "kilometer"
-  ) => void;
-  handleClearFiltersClick: () => void;
-  accentColorCssValue: string;
-  closeButtonRef: React.Ref<HTMLButtonElement>;
-}
-
-const FilterModal = (props: FilterModalProps) => {
-  const {
-    showFilterModal,
-    showOpenNowOption,
-    isOpenNowSelected,
-    showDistanceOptions,
-    selectedDistanceOption,
-    handleCloseModalClick,
-    handleOpenNowClick,
-    handleDistanceClick,
-    handleClearFiltersClick,
-    accentColorCssValue,
-    closeButtonRef,
-  } = props;
-  const { t } = useTranslation();
-  const popupRef = React.useRef<HTMLDivElement>(null);
-
-  return showFilterModal ? (
-    <div
-      id="locator-filter-modal"
-      role="dialog"
-      aria-modal="true"
-      aria-labelledby="locator-filter-modal-title"
-      className="absolute md:top-4 -top-20 z-50 md:w-80 w-full flex flex-col bg-white md:left-full md:ml-2 rounded-md shadow-lg max-h-[calc(100%-2rem)]"
-      style={
-        {
-          "--locator-filter-accent-color": accentColorCssValue,
-        } as React.CSSProperties
-      }
-      ref={popupRef}
-    >
-      <div className="inline-flex justify-between items-center px-6 py-4 gap-4">
-        <Body className="font-bold" id="locator-filter-modal-title">
-          {t("refineYourSearch", "Refine Your Search")}
-        </Body>
-        <button
-          ref={closeButtonRef}
-          style={{ color: accentColorCssValue }}
-          onClick={handleCloseModalClick}
-          aria-label={t("close", "Close")}
-        >
-          <FaTimes />
-        </button>
-      </div>
-      <div className="px-6 border-b border-gray-300">
-        <AppliedFilters
-          hiddenFields={[LOCATION_FIELD, COUNTRY_CODE_FIELD]}
-          customCssClasses={{
-            removableFilter: "text-md font-normal",
-            clearAllButton: "hidden",
-          }}
-        />
-      </div>
-      <div className="flex flex-col p-6 gap-6 overflow-y-auto">
-        <div className="flex flex-col gap-8">
-          {showOpenNowOption && (
-            <OpenNowFilter
-              isSelected={isOpenNowSelected}
-              onChange={handleOpenNowClick}
-            />
-          )}
-          {showDistanceOptions && (
-            <DistanceFilter
-              onChange={handleDistanceClick}
-              selectedDistanceOption={selectedDistanceOption}
-              accentColorCssValue={accentColorCssValue}
-            />
-          )}
-          <Facets
-            customCssClasses={{
-              divider: "bg-white",
-              titleLabel: "font-bold text-md font-body-fontFamily",
-              optionInput:
-                "h-4 w-4 [accent-color:var(--locator-filter-accent-color)]",
-              optionLabel: "text-md font-body-fontFamily font-body-fontWeight",
-              option: "space-x-4 font-body-fontFamily",
-            }}
-          />
-        </div>
-      </div>
-      <div className="border-y border-gray-300 justify-center align-middle">
-        <button
-          className="w-full py-4 text-center font-bold font-body-fontFamily text-body-fontSize"
-          style={{ color: accentColorCssValue }}
-          onClick={handleClearFiltersClick}
-        >
-          {t("clearAll", "Clear All")}
-        </button>
-      </div>
-    </div>
-  ) : null;
-};
-
-interface OpenNowFilterProps {
-  isSelected: boolean;
-  onChange: (selected: boolean) => void;
-}
-
-const OpenNowFilter = (props: OpenNowFilterProps) => {
-  const { isSelected, onChange } = props;
-  const { t } = useTranslation();
-  const { isExpanded, getToggleProps, getCollapseProps } = useCollapse({
-    defaultExpanded: true,
-  });
-  const iconClassName = isExpanded
-    ? "w-3 text-gray-400"
-    : "w-3 text-gray-400 transform rotate-180";
-
-  const openNowCheckBoxId = "openNowCheckBox";
-  return (
-    <div className="flex flex-col gap-4">
-      <button
-        className="w-full flex justify-between items-center font-bold font-body-fontFamily text-body-fontSize"
-        {...getToggleProps()}
-      >
-        {t("hours", "Hours")}
-        <FaChevronUp className={iconClassName} />
-      </button>
-      <div className="flex flex-row gap-1" {...getCollapseProps()}>
-        <div className="inline-flex items-center gap-4">
-          <input
-            type="checkbox"
-            id={openNowCheckBoxId}
-            checked={isSelected}
-            className={
-              "w-4 h-4 form-checkbox cursor-pointer border border-gray-300" +
-              " rounded-sm text-primary focus:ring-primary [accent-color:var(--locator-filter-accent-color)]"
-            }
-            onChange={() => onChange(!isSelected)}
-          />
-          <label htmlFor={openNowCheckBoxId}>
-            <Body>{t("openNow", "Open Now")}</Body>
-          </label>
-        </div>
-      </div>
-    </div>
-  );
-};
-
-interface DistanceFilterProps {
-  onChange: (distance: number, unit: "mile" | "kilometer") => void;
-  selectedDistanceOption: number | null;
-  accentColorCssValue: string;
-}
-
-const DistanceFilter = (props: DistanceFilterProps) => {
-  const { selectedDistanceOption, onChange, accentColorCssValue } = props;
-  const { t, i18n } = useTranslation();
-  const { isExpanded, getToggleProps, getCollapseProps } = useCollapse({
-    defaultExpanded: true,
-  });
-  const iconClassName = isExpanded
-    ? "w-3 text-gray-400"
-    : "w-3 text-gray-400 transform rotate-180";
-  const distanceOptions = [5, 10, 25, 50];
-  const unit = getPreferredDistanceUnit(i18n.language);
-
-  return (
-    <div className="flex flex-col gap-4">
-      <button
-        className="w-full flex justify-between items-center font-bold font-body-fontFamily text-body-fontSize"
-        {...getToggleProps()}
-      >
-        {t("distance", "Distance")}
-        <FaChevronUp className={iconClassName} />
-      </button>
-      <div {...getCollapseProps()}>
-        {distanceOptions.map((distanceOption) => (
-          <div
-            className="flex flex-row gap-4 items-center"
-            id={`distanceOption-${distanceOption}`}
-            key={distanceOption}
-          >
-            <button
-              className="inline-flex bg-white"
-              onClick={() => onChange(distanceOption, unit)}
-              aria-label={`${t("selectDistanceLessThan", "Select distance less than")} ${distanceOption} ${translateDistanceUnit(t, unit, distanceOption)}`}
-            >
-              <div style={{ color: accentColorCssValue }}>
-                {selectedDistanceOption === distanceOption ? (
-                  <FaDotCircle />
-                ) : (
-                  <FaRegCircle />
-                )}
-              </div>
-            </button>
-            <Body className="inline-flex">
-              {`< ${distanceOption} ${translateDistanceUnit(t, unit, distanceOption)}`}
-            </Body>
-          </div>
-        ))}
-      </div>
-    </div>
-  );
-};
-
-const getMapboxMapPadding = (divElement: HTMLDivElement | null) => {
-  if (!divElement) {
-    return 50;
-  }
-
-  const { width, height } = divElement.getBoundingClientRect();
-  const mapVerticalPadding = Math.max(50, height * 0.2);
-  const mapHorizontalPadding = Math.max(50, width * 0.2);
-  return {
-    top: mapVerticalPadding,
-    bottom: mapVerticalPadding,
-    left: mapHorizontalPadding,
-    right: mapHorizontalPadding,
-  };
-};
-
-const parseMapStartingLocation = (mapStartingLocation: {
-  latitude: string;
-  longitude: string;
-}): Coordinate => {
-  const lat = parseFloat(mapStartingLocation.latitude);
-  const lng = parseFloat(mapStartingLocation.longitude);
-
-  let err = [];
-  if (isNaN(lat) || lat < -90 || lat > 90) {
-    err.push("Latitude must be a number between -90 and 90.");
-  }
-  if (isNaN(lng) || lng < -180 || lng > 180) {
-    err.push("Longitude must be a number between -180 and 180.");
-  }
-  if (err.length) {
-    throw new Error(err.join("\n"));
-  }
-
-  return {
-    latitude: lat,
-    longitude: lng,
-  };
-};
-
-/**
- * Returns true if the given filter is a "near" filter on the builtin.location field; otherwise,
- * returns false.
- */
-const isLocationNearFilter = (filter: SelectableStaticFilter) =>
-  filter.filter.kind === "fieldValue" &&
-  filter.filter.fieldId === LOCATION_FIELD &&
-  filter.filter.matcher === Matcher.Near;
-
-/**
- * Returns true if the given filter is an "open at" filter on the builtin.hours field; otherwise,
- * returns false.
- */
-const isOpenNowFilter = (filter: SelectableStaticFilter) =>
-  filter.filter.kind === "fieldValue" &&
-  filter.filter.fieldId === HOURS_FIELD &&
-  filter.filter.matcher === Matcher.OpenAt;
-
-/**
- * Builds a "near" static filter on the builtin.location field from a previous near filter
- * value, with optional overrides for display name and radius
- */
-function buildNearLocationFilterFromPrevious(
-  previousValue: NearFilterValue,
-  displayName?: string,
-  radius?: number
-): SelectableStaticFilter {
-  return {
-    selected: true,
-    displayName,
-    filter: {
-      kind: "fieldValue",
-      fieldId: LOCATION_FIELD,
-      value: {
-        ...previousValue,
-        radius: radius ?? previousValue.radius,
-      },
-      matcher: Matcher.Near,
-    },
-  };
-}
-
-/**
- * Builds a "near" static filter on the builtin.location field from given coordinates, with
- * optional radius and display name.
- */
-function buildNearLocationFilterFromCoords(
-  lat: number,
-  lng: number,
-  radius: number,
-  displayName?: string
-): SelectableStaticFilter {
-  return {
-    selected: true,
-    displayName,
-    filter: {
-      kind: "fieldValue",
-      fieldId: LOCATION_FIELD,
-      value: {
-        lat,
-        lng,
-        radius,
-      },
-      matcher: Matcher.Near,
-    },
-  };
-}
-
-/**
- * Builds an "equals" static filter on the builtin.location field from a previous equals filter,
- * with a new display name.
- */
-function buildEqualsLocationFilter(
-  filter: FieldValueFilter,
-  newDisplayName: string
-): SelectableStaticFilter {
-  return {
-    displayName: newDisplayName,
-    selected: true,
-    filter: {
-      kind: "fieldValue",
-      fieldId: filter.fieldId,
-      value: filter.value,
-      matcher: Matcher.Equals,
-    },
-  };
-}
-
-/**
- * Helper function to iterate through a list of static filters and update all near filters on the
- * location field to have the new radius.
- */
-function updateRadiusInNearFiltersOnLocationField(
-  filters: SelectableStaticFilter[],
-  newRadius: number
-): SelectableStaticFilter[] {
-  return filters.map((filter) => {
-    if (isLocationNearFilter(filter)) {
-      const previousFilter = filter.filter as FieldValueStaticFilter;
-      const previousValue = previousFilter.value as NearFilterValue;
-      return {
-        ...filter,
-        filter: {
-          ...previousFilter,
-          value: {
-            ...previousValue,
-            radius: newRadius,
-          },
-        },
-      } as SelectableStaticFilter;
-    }
-    return filter;
-  });
-}
-
-/**
- * Helper function to iterate through a list of static filters and set the selected field to
- * false on any Open Now filters.
- */
-function deselectOpenNowFilters(
-  filters: SelectableStaticFilter[]
-): SelectableStaticFilter[] {
-  return filters.map((filter) => {
-    if (isOpenNowFilter(filter)) {
-      return {
-        ...filter,
-        selected: false,
-      };
-    }
-    return filter;
-  });
-}
-
-/** Checks whether a given lat and lng are valid coordinates */
-function areValidCoordinates(lat: number, lng: number): boolean {
-  return (
-    !isNaN(lat) &&
-    !isNaN(lng) &&
-    lat >= -90 &&
-    lat <= 90 &&
-    lng >= -180 &&
-    lng <= 180
-  );
-}
